requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
duckdb>=1.1.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyvis>=0.3.2
networkx>=3.2
scikit-learn>=1.5.0
scipy>=1.12.0
user-agents>=2.2.0
tldextract>=5.1.0
python-dotenv>=1.0.0
rich>=13.7.0
